Blog_to_Podcast/Blog_to_Podcast-main/blog_to_podcast/requirements.txt
streamlit
openai
requests
python-dotenv
elevenlabs
beautifulsoup4
lxml
html5lib
